requirements.txt
# 核心依赖
streamlit==1.28.0
pyyaml==6.0
requests==2.31.0
httpx==0.24.1
openai==0.28.0

# 数据处理
pandas==2.0.3
numpy==1.24.3

# 向量检索
faiss-cpu==1.7.4
scikit-learn==1.3.0

# 文档处理
python-docx==0.8.11
PyPDF2==3.0.1
chardet==5.1.0

# 模型服务
xinference==0.5.0

# 工具库
tqdm==4.65.0


# xinference
# 建立xinference的虚拟环境
# conda create -n xinference python=3.10
# conda activate xinference

# 安装xinference
# pip install xinference

